split specs module into three submodules
