+++ cs251tk/specs/__init__.py
@@ -1,3 +1,3 @@
-from .specs import load_all_specs
-from .specs import load_some_specs
-from .specs import get_filenames
+from .load import load_all_specs
+from .load import load_some_specs
+from .util import get_filenames
