+++ cs251tk/specs/util.py
@@ -1,0 +1,3 @@
+def get_filenames(spec):
+    """returns the list of files from an assignment spec"""
+    return [file['filename'] for file in spec['files']]
